refactor: relocate /internal/constant -> /common/constant
